requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
google-genai==1.14.0
supabase==2.11.0
pydantic-settings==2.7.1
python-multipart==0.0.20
